+++ src/pine/data_types.clj
@@ -7,3 +7,6 @@
 
 (defn number [x]
   (t :number x))
+
+(defn symbol [x]
+  (t :symbol x))