+++ src/pine/eval.clj
@@ -35,6 +35,10 @@
      select-all
      " FROM")))
 
+(defn- remove-symbols [vs]
+  "Remove symbols from a vector of values"
+  (filter #(not (= (:type %) :symbol)) vs))
+
 (defn build-select-clause [state]
   (let [{:keys [tables columns limit where aliases]} state
         from         (let [{a :alias} (first tables)
@@ -46,15 +50,15 @@
                        (str "WHERE "
                             (clojure.string/join " AND "
                                                  (for [[a col op value] where]
-                                                   (if (= op "in")
+                                                   (if (= op "IN")
                                                      (str (q a col) " IN (" (clojure.string/join ", " (repeat (count value) "?"))  ")")
-                                                     (str (q a col) " " op " ?"))))))
+                                                     (str (q a col) " " op " " (if (= (:type value) :symbol) (:value value) "?")))))))
         limit (str "LIMIT " (or limit 250))
         query (clojure.string/join " " (filter some? [select from join where-clause limit]))
         params (when (not-empty where)
                  (->> where
                       (map (fn [[a col op value]] (if (coll? value) value [value])))
-                      ;; TODO: filter out the symbols e.g. null, etc
+                      remove-symbols
                       flatten))]
 
     {:query query :params params}))