+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file. This change
 log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+
+- Support for `NULL`:
+
+```
+company | name is null
+company | name = null
+```
 
 
 ## [0.8.1] - 2024-07-30